restructuring tools.py
