+++ dbitx_funcs/__init__.py
@@ -4,9 +4,3 @@ from .readwrite import *
 from . import calculations as calc
 from . import images as im
 from . import utils as utils
-
-
-
-
-
-
