+++ dbitx_funcs/calculations/__init__.py
@@ -1,10 +1,3 @@
 from ._calc import *
 from .contours import *
 from .clustering import *
-
-
-
-
-
-
-
