+++ dbitx_funcs/tools/__init__.py
@@ -1,4 +1,4 @@
 from .tools import *
 from .go import *
-
-
+from .adata import *
+from .preprocessing import *
